マップ画面に管理者・開発者用ハンバーガーメニューを追加
