+++ client/src/routes/map/index.tsx
@@ -1,15 +1,11 @@
 import { createFileRoute, redirect } from "@tanstack/react-router";
 import MapComponent from "@/components/map/map";
-import { isAdminUser } from "@/lib/adminRole";
 import { getAccessToken } from "@/lib/apiClient";
 
 export const Route = createFileRoute("/map/")({
   beforeLoad: () => {
     if (!getAccessToken()) {
       throw redirect({ to: "/login" });
-    }
-    if (getAccessToken() && isAdminUser()) {
-      throw redirect({ to: "/dashboard" });
     }
   },
   component: MapComponent,
